Support date column and from and to dict
