+++ tests/open_alchemy/utility_base/test_to_dict.py
@@ -9,14 +9,8 @@ from open_alchemy import exceptions
 from open_alchemy import utility_base
 
 
-def __init__(self, **kwargs):
-    """Construct."""
-    for name, value in kwargs.items():
-        setattr(self, name, value)
-
-
 @pytest.mark.utility_base
-def test_to_dict_no_schema():
+def test_to_dict_no_schema(__init__):
     """
     GIVEN class that derives from UtilityBase but does not define _schema
     WHEN to_dict is called
@@ -30,7 +24,7 @@ def test_to_dict_no_schema():
 
 
 @pytest.mark.utility_base
-def test_to_dict_no_properties():
+def test_to_dict_no_properties(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema without properties
     WHEN to_dict is called
@@ -46,7 +40,7 @@ def test_to_dict_no_properties():
 
 
 @pytest.mark.utility_base
-def test_to_dict_no_type():
+def test_to_dict_no_type(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema with a property without a
         type
@@ -90,7 +84,7 @@ def test_to_dict_no_type():
     ],
 )
 @pytest.mark.utility_base
-def test_to_dict_simple_type(schema, init_args, expected_dict):
+def test_to_dict_simple_type(schema, init_args, expected_dict, __init__):
     """
     GIVEN class that derives from UtilityBase with a given schema with properties that
         are not objects and expected object
@@ -109,7 +103,7 @@ def test_to_dict_simple_type(schema, init_args, expected_dict):
 
 @pytest.mark.parametrize("init_kwargs", [{}, {"key": None}], ids=["undefined", "none"])
 @pytest.mark.utility_base
-def test_to_dict_object_none(init_kwargs):
+def test_to_dict_object_none(init_kwargs, __init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an object property and
         init args
@@ -129,7 +123,7 @@ def test_to_dict_object_none(init_kwargs):
 
 
 @pytest.mark.utility_base
-def test_to_dict_object():
+def test_to_dict_object(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an object property
         that has a mock model
@@ -151,7 +145,7 @@ def test_to_dict_object():
 
 @pytest.mark.parametrize("init_kwargs", [{}, {"key": None}], ids=["undefined", "none"])
 @pytest.mark.utility_base
-def test_to_dict_array_none(init_kwargs):
+def test_to_dict_array_none(init_kwargs, __init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an array property and
         init args
@@ -176,7 +170,7 @@ def test_to_dict_array_none(init_kwargs):
 
 
 @pytest.mark.utility_base
-def test_to_dict_array_empty():
+def test_to_dict_array_empty(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an array property
         that is empty
@@ -201,7 +195,7 @@ def test_to_dict_array_empty():
 
 
 @pytest.mark.utility_base
-def test_to_dict_array_single():
+def test_to_dict_array_single(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an array property
         that has a single mock model
@@ -228,7 +222,7 @@ def test_to_dict_array_single():
 
 
 @pytest.mark.utility_base
-def test_to_dict_array_multiple():
+def test_to_dict_array_multiple(__init__):
     """
     GIVEN class that derives from UtilityBase with a schema with an array property
         that has multiple mock models
@@ -256,396 +250,6 @@ def test_to_dict_array_multiple():
             map(lambda mock_model: mock_model.to_dict.return_value, mock_models)
         )
     }
-
-
-@pytest.mark.utility_base
-def test_to_dict_malformed_dictionary():
-    """
-    GIVEN class that derives from UtilityBase and schema
-    WHEN from_dict is called with a dictionary that does not satisfy the schema
-    THEN MalformedModelDictionaryError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {"key": {"type": "integer"}},
-                "required": ["key"],
-            },
-            "__init__": __init__,
-        },
-    )
-
-    with pytest.raises(exceptions.MalformedModelDictionaryError):
-        model.from_dict(**{})
-
-
-@pytest.mark.utility_base
-def test_from_dict_argument_not_in_properties():
-    """
-    GIVEN dictionary with a key which is not a property in the schema
-    WHEN from_dict is called with the dictionary
-    THEN MalformedModelDictionaryError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {"_schema": {"properties": {}}, "__init__": __init__},
-    )
-
-    with pytest.raises(exceptions.MalformedModelDictionaryError):
-        model.from_dict(**{"key": "value"})
-
-
-@pytest.mark.utility_base
-def test_from_dict_read_only():
-    """
-    GIVEN schema with readOnly and dictionary
-    WHEN from_dict is called with the dictionary
-    THEN MalformedModelDictionaryError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {"properties": {"key": {"readOnly": True, "type": "string"}}},
-            "__init__": __init__,
-        },
-    )
-
-    with pytest.raises(exceptions.MalformedModelDictionaryError):
-        model.from_dict(**{"key": "value"})
-
-
-@pytest.mark.utility_base
-def test_from_dict_read_only_false():
-    """
-    GIVEN schema with readOnly and dictionary
-    WHEN from_dict is called with the dictionary
-    THEN MalformedModelDictionaryError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {"properties": {"key": {"readOnly": False, "type": "string"}}},
-            "__init__": __init__,
-        },
-    )
-
-    model.from_dict(**{"key": "value"})
-
-
-@pytest.mark.utility_base
-def test_from_dict_no_type_schema():
-    """
-    GIVEN model with a schema with a property without a type
-    WHEN from_dict is called with the dictionary
-    THEN MissingTypeError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {"_schema": {"properties": {"key": {}}}, "__init__": __init__},
-    )
-
-    with pytest.raises(exceptions.TypeMissingError):
-        model.from_dict(**{"key": "value"})
-
-
-@pytest.mark.parametrize(
-    "schema, dictionary",
-    [
-        ({"properties": {"key_1": {"type": "integer"}}}, {}),
-        ({"properties": {"key_1": {"type": "integer"}}}, {"key_1": 1}),
-        (
-            {"properties": {"key_1": {"type": "integer"}}, "required": ["key_1"]},
-            {"key_1": 1},
-        ),
-        (
-            {
-                "properties": {
-                    "key_1": {"type": "integer"},
-                    "key_2": {"type": "integer"},
-                }
-            },
-            {"key_1": 11, "key_2": 12},
-        ),
-    ],
-    ids=[
-        "single not required not given",
-        "single not required given",
-        "single required given",
-        "multiple",
-    ],
-)
-@pytest.mark.utility_base
-def test_from_dict(schema, dictionary):
-    """
-    GIVEN schema and dictionary to use for construction
-    WHEN model is defined with the schema and constructed with from_dict
-    THEN the instance has the properties from the dictionary.
-    """
-    model = type(
-        "model", (utility_base.UtilityBase,), {"_schema": schema, "__init__": __init__}
-    )
-
-    instance = model.from_dict(**dictionary)
-
-    for key, value in dictionary.items():
-        assert getattr(instance, key) == value
-
-
-@pytest.mark.utility_base
-def test_from_dict_date_time():
-    """
-    GIVEN schema with string type and date-time format
-    WHEN model is defined and constructed with from_dict
-    THEN the instance has a datetime.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {"key": {"type": "string", "format": "date-time"}}
-            },
-            "__init__": __init__,
-        },
-    )
-    dictionary = {"key": "2000-01-01T01:01:01"}
-
-    instance = model.from_dict(**dictionary)
-
-    assert instance.key == datetime.datetime(  # pylint: disable=no-member
-        year=2000, month=1, day=1, hour=1, minute=1, second=1
-    )
-
-
-@pytest.mark.utility_base
-def test_from_dict_object_de_ref_missing():
-    """
-    GIVEN schema with object without x-de-$ref
-    WHEN model is defined and constructed with from_dict
-    THEN MalformedSchemaError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {"_schema": {"properties": {"key": {"type": "object"}}}, "__init__": __init__},
-    )
-
-    with pytest.raises(exceptions.MalformedSchemaError):
-        model.from_dict(**{"key": {"obj_key": "obj value"}})
-
-
-@pytest.mark.utility_base
-def test_from_dict_object_model_undefined():
-    """
-    GIVEN schema with object which references a model that has not been defined
-    WHEN from_dict is called
-    THEN SchemaNotFoundError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {"key": {"type": "object", "x-de-$ref": "RefModel"}}
-            },
-            "__init__": __init__,
-        },
-    )
-
-    with pytest.raises(exceptions.SchemaNotFoundError):
-        model.from_dict(**{"key": {"obj_key": "obj value"}})
-
-
-@pytest.mark.utility_base
-def test_from_dict_object_from_dict_call(mocked_facades_models):
-    """
-    GIVEN schema with object which references a model that has been mocked and
-        dictionary
-    WHEN from_dict is called with the dictionary
-    THEN from_dict on the mocked model is called with the portion of the dictionary
-        for that model.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {"key": {"type": "object", "x-de-$ref": "RefModel"}}
-            },
-            "__init__": __init__,
-        },
-    )
-
-    model.from_dict(**{"key": {"obj_key": "obj value"}})
-
-    mocked_facades_models.get_model.assert_called_once_with(name="RefModel")
-    mocked_facades_models.get_model.return_value.from_dict.assert_called_once_with(
-        **{"obj_key": "obj value"}
-    )
-
-
-@pytest.mark.utility_base
-def test_from_dict_object_return(mocked_facades_models):
-    """
-    GIVEN schema with object which references a model that has been mocked and
-        dictionary
-    WHEN from_dict is called with the dictionary
-    THEN the from_dict on the mocked model return value is bound to the model instance.
-        for that model.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {"key": {"type": "object", "x-de-$ref": "RefModel"}}
-            },
-            "__init__": __init__,
-        },
-    )
-
-    instance = model.from_dict(**{"key": {"obj_key": "obj value"}})
-
-    mocked_facades_models.get_model.assert_called_once_with(name="RefModel")
-    assert (
-        instance.key  # pylint: disable=no-member
-        == mocked_facades_models.get_model.return_value.from_dict.return_value
-    )
-
-
-@pytest.mark.utility_base
-def test_from_dict_array_no_items():
-    """
-    GIVEN schema with array which does not define items
-    WHEN from_dict is called with the dictionary
-    THEN MalformedSchemaError is raised.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {"_schema": {"properties": {"key": {"type": "array"}}}, "__init__": __init__},
-    )
-
-    with pytest.raises(exceptions.MalformedSchemaError):
-        model.from_dict(**{"key": []})
-
-
-@pytest.mark.utility_base
-def test_from_dict_array_empty_from_dict(mocked_facades_models):
-    """
-    GIVEN schema with array which references a model that has been mocked and
-        dictionary with an empty array
-    WHEN from_dict is called with the dictionary
-    THEN from_dict on the mocked model is not called and the value is set to an empty
-        list.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {
-                    "key": {
-                        "type": "array",
-                        "items": {"type": "object", "x-de-$ref": "RefModel"},
-                    }
-                }
-            },
-            "__init__": __init__,
-        },
-    )
-
-    instance = model.from_dict(**{"key": []})
-
-    mocked_facades_models.get_model.return_value.from_dict.assert_not_called()
-    assert instance.key == []  # pylint: disable=no-member
-
-
-@pytest.mark.utility_base
-def test_from_dict_array_single_from_dict(mocked_facades_models):
-    """
-    GIVEN schema with array which references a model that has been mocked and
-        dictionary with a single item array
-    WHEN from_dict is called with the dictionary
-    THEN from_dict on the mocked model is called with the portion of the dictionary
-        for that model and the value is set to a list of the model from_dict return
-        value.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {
-                    "key": {
-                        "type": "array",
-                        "items": {"type": "object", "x-de-$ref": "RefModel"},
-                    }
-                }
-            },
-            "__init__": __init__,
-        },
-    )
-
-    instance = model.from_dict(**{"key": [{"obj_key": "obj value"}]})
-
-    mocked_facades_models.get_model.return_value.from_dict.assert_called_once_with(
-        **{"obj_key": "obj value"}
-    )
-    mocked_facades_models.get_model.assert_called_once_with(name="RefModel")
-    assert instance.key == [  # pylint: disable=no-member
-        mocked_facades_models.get_model.return_value.from_dict.return_value
-    ]
-
-
-@pytest.mark.utility_base
-def test_from_dict_array_multiple_from_dict(mocked_facades_models):
-    """
-    GIVEN schema with array which references a model that has been mocked and
-        dictionary with a multiple item array
-    WHEN from_dict is called with the dictionary
-    THEN from_dict on the mocked model is called with the portion of the dictionary
-        for that model and the value is set to a list of the model from_dict return
-        value.
-    """
-    model = type(
-        "model",
-        (utility_base.UtilityBase,),
-        {
-            "_schema": {
-                "properties": {
-                    "key": {
-                        "type": "array",
-                        "items": {"type": "object", "x-de-$ref": "RefModel"},
-                    }
-                }
-            },
-            "__init__": __init__,
-        },
-    )
-
-    instance = model.from_dict(
-        **{"key": [{"obj_key_1": "obj value 1"}, {"obj_key_2": "obj value 2"}]}
-    )
-
-    mocked_facades_models.get_model.assert_called_once_with(name="RefModel")
-    assert mocked_facades_models.get_model.return_value.from_dict.call_count == 2
-    mocked_facades_models.get_model.return_value.from_dict.assert_any_call(
-        **{"obj_key_1": "obj value 1"}
-    )
-    mocked_facades_models.get_model.return_value.from_dict.assert_any_call(
-        **{"obj_key_2": "obj value 2"}
-    )
-    assert instance.key == [  # pylint: disable=no-member
-        mocked_facades_models.get_model.return_value.from_dict.return_value,
-        mocked_facades_models.get_model.return_value.from_dict.return_value,
-    ]
 
 
 class TestObjectToDictRelationship:
@@ -832,21 +436,34 @@ class TestToDictProperty:
         assert returned_value == value
 
     @staticmethod
+    @pytest.mark.parametrize(
+        "format_, value, expected_value",
+        [
+            ("date", datetime.date(year=2000, month=1, day=1), "2000-01-01"),
+            (
+                "date-time",
+                datetime.datetime(
+                    year=2000, month=1, day=1, hour=1, minute=1, second=1
+                ),
+                "2000-01-01T01:01:01",
+            ),
+        ],
+        ids=["date", "date-time"],
+    )
     @pytest.mark.utility_base
-    def test_date_time():
+    def test_string_format(format_, value, expected_value):
         """
-        GIVEN spec that is a date time and value
-        WHEN _to_dict_property is called with the spec and value
-        THEN value is returned as a string.
+        GIVEN spec with a string type and given format
+        WHEN _to_dict_property is called with the spec and given value
+        THEN the given expected value is returned as a string.
         """
-        spec = {"type": "string", "format": "date-time"}
-        value = datetime.datetime(year=2000, month=1, day=1, hour=1, minute=1, second=1)
+        spec = {"type": "string", "format": format_}
 
         returned_value = utility_base.UtilityBase._to_dict_property(
             value, spec=spec, name="name 1"
         )
 
-        assert returned_value == "2000-01-01T01:01:01"
+        assert returned_value == expected_value
 
     @staticmethod
     @pytest.mark.utility_base
